Move macro to its own file
